html/semantics/embedded-content-0/the-canvas-element/size.attributes.removed.html

<!DOCTYPE html>
<title>Canvas test: size.attributes.removed</title>
<meta name="author" content="Philip Taylor">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../../../common/canvas-tests.js"></script>
<link rel="stylesheet" href="../../../../../common/canvas-tests.css">
<body class="show_output">

<h1><a href="index.size.html">size</a>.<a href="index.size.attributes.html">attributes</a>.removed</h1>
<p class="desc">Removing content attributes reverts to default size</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.size.missing">size.missing</a>
</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="120" height="60"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="size.attributes.removed.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Removing content attributes reverts to default size");
_addTest(function(canvas, ctx) {

_assertSame(canvas.width, 120, "canvas.width", "120");
canvas.removeAttribute('width');
_assertSame(canvas.width, 300, "canvas.width", "300");


});
</script>
